move dummy to configs for production
