--- src/view/home/heroes/dummy.ts
@@ -1,13 +1,0 @@
-import { IMG_SERABI } from '../../../configs/images'
-
-const args = {
-  img: IMG_SERABI,
-  button: {
-    name: 'beli sekarang!',
-    url: '#beli',
-  },
-  description:
-    'Lorem ipsum dolor sit amet, qui minim labore adipisicing minim sint cillum sint consectetur cupidatat.',
-  title: 'serabi ngampin bu yuni',
-}
-export default args
