+++ src/view/home/page/index.ts
@@ -3,20 +3,21 @@ import contact, { type TArgs as TContact } from '../contact'
 import Heroes, { type TArgs as THeroes } from '../heroes'
 
 export type TArgs = {
-  heroesArgs: THeroes
   Menu: string
   aboutArgs: TAbout
   contactArgs: TContact
   telp: number
+  heroesArgs: THeroes
 }
 
 export default function HomePage({
-  heroesArgs,
   Menu,
   aboutArgs,
   contactArgs,
+  heroesArgs,
 }: TArgs) {
   return html`
-      ${Heroes(heroesArgs) + Menu}
-      <div class="md:flex">${About(aboutArgs) + contact(contactArgs)}</div>
-    `}
+    ${Heroes(heroesArgs) + Menu}
+    <div class="md:flex">${About(aboutArgs) + contact(contactArgs)}</div>
+  `
+}
